v0.24
Created a cancel button in W_TowerSelection to remove the widget from the screen. When a tower is placed or CancelButton is pressed, the widget is removed from screen and the tile is unselected.

Fixed issue when clicking a tile while W_TowerSelection exists, the widget is not removed and needs to be pressed for however many tiles were selected. Example:
User selects three tiles and then presses cancel. Three instances of W_TowerSelection were added to viewport, but only one is removed.